requirements.txt
alembic==1.0.0
argcomplete==1.10.0
beautifulsoup4==4.8.0
certifi==2020.4.5.1
chardet==3.0.4
click==6.7
docx2txt==0.8
EbookLib==0.17.1
extract-msg==0.23.1
Flask==1.0.2
Flask-Migrate==2.2.1
Flask-Script==2.0.6
Flask-SQLAlchemy==2.4.1
gunicorn==19.9.0
idna==2.9
IMAPClient==2.1.0
itsdangerous==0.24
Jinja2==2.10
lxml==4.5.1
Mako==1.1.2
MarkupSafe==1.0
olefile==0.46
pdfminer.six==20181108
Pillow==7.1.2
psycopg2==2.8.4
psycopg2-binary==2.8.5
pycryptodome==3.9.7
PyPDF2==1.26.0
python-dateutil==2.7.3
python-editor==1.0.3
python-pptx==0.6.18
pytz==2020.1
requests==2.23.0
six==1.11.0
sortedcontainers==2.1.0
soupsieve==2.0.1
SpeechRecognition==3.8.1
SQLAlchemy==1.3.17
textract==1.6.3
tika==1.24
tzlocal==1.5.1
urllib3==1.25.9
Werkzeug==0.14.1
xlrd==1.2.0
XlsxWriter==1.2.8
